% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{Colorplot}
\alias{Colorplot}
\title{Sequence optimize}
\usage{
Colorplot(
  seuInt,
  save.path,
  name = "Int_cluster_",
  cluster = NULL,
  cols_cluster = c("#FD7446", "#709AE1", "#31A354", "#9EDAE5", "#DE9ED6", "#BCBD22",
    "#CE6DBD", "#DADAEB", "yellow", "#FF9896", "#91D1C2", "#C7E9C0", "#6B6ECF",
    "#7B4173")
)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}

\item{model}{the model to be used. The default is set as 'linear' to use the linear model.}

\item{option}{a list specifies the options for model solving, currently supporting 'mle_solver' only.}
}
\value{
A list including the information of the fitted model
}
\description{
Do sequence optimize
}
