% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{Gene_mat_generate}
\alias{Gene_mat_generate}
\title{Sequence optimize}
\usage{
Gene_mat_generate(seuInt, markers, k = 20, save.path, name = NULL)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}

\item{model}{the model to be used. The default is set as 'linear' to use the linear model.}

\item{option}{a list specifies the options for model solving, currently supporting 'mle_solver' only.}
}
\value{
A list including the information of the fitted model
}
\description{
Do sequence optimize
}
