% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{plot_gene_by_gene}
\alias{plot_gene_by_gene}
\title{Sequence optimize}
\usage{
plot_gene_by_gene(
  gene1,
  gene2,
  i,
  j,
  p_adj,
  expr_mat,
  neighbor_mat,
  neighbor_mat_list,
  cell_by = seuInt$label,
  smooth_by = seuInt$sample,
  facet_by = seuInt$batch,
  zero_rm = F
)
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}

\item{model}{the model to be used. The default is set as 'linear' to use the linear model.}

\item{option}{a list specifies the options for model solving, currently supporting 'mle_solver' only.}
}
\value{
A list including the information of the fitted model
}
\description{
Do sequence optimize
}
